Update Website php to use requestedFloor
This merge request is being put into use requseted floor when the user
submits a request for the elevator to change floors from the website.
the orignal impementation used current floor so even if the elevator
didn't change floors when the website will start to display the floor
the elevator is on it could be wrong since it never gets changed by the
controller if it doesn't run.